shopai_full_code/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
firebase-admin==6.5.0
google-cloud-storage==2.16.0
requests==2.32.3
python-dotenv==1.0.1
